- Fixed a memory leak in uniform lifting and type conversion. - Fixed type conversion with scalars (it now seems to work and return the correct types). - Changed ShLib stuff to use "float" constants to avoid excessive type conversions (now that they work ;). Eventually we should provide versions for float, double, int, etc.
git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/branches/0.8@3806 afdca40c-03d6-0310-8ede-e9f093b21075

+++ src/sh/ShLibMisc.hpp
@@ -42,7 +42,7 @@ namespace SH {
 template<int M, int N, typename T> 
 ShGeneric<M, T> cast(const ShGeneric<N, T>& a);
 template<int M> 
-ShGeneric<M, double> cast(double a); // @todo type do a cpp type -> value type map
+ShGeneric<M, float> cast(float a); // @todo type do a cpp type -> value type map
 
 /** Fill Casting.
  * Casts ShGeneric<N, T> to ShGeneric<M, T>.
@@ -52,7 +52,7 @@ ShGeneric<M, double> cast(double a); // @todo type do a cpp type -> value type m
 template<int M, int N, typename T> 
 ShGeneric<M, T> fillcast(const ShGeneric<N, T>& a);
 template<int M> 
-ShGeneric<M, double> fillcast(double a); // @todo type do a cpp type -> value type map
+ShGeneric<M, float> fillcast(float a); // @todo type do a cpp type -> value type map
 
 /** Join two tuples 
  * Creates an M+N tuple with components of a first then b.
